Move `ActorPF2e#findStackableItem` to `ActorInventory` and use `Remeda.equals` to determine stackability
